feat: add clone environment api
