+++ pkg/apis/environment/extension_view.go
@@ -3,6 +3,7 @@ package environment
 import (
 	"github.com/seal-io/walrus/pkg/dao/model"
 	"github.com/seal-io/walrus/pkg/dao/types"
+	"github.com/seal-io/walrus/pkg/dao/types/object"
 )
 
 type (
@@ -27,3 +28,19 @@ type (
 		Edges    []GraphEdge   `json:"edges"`
 	}
 )
+
+type (
+	RouteCloneEnvironmentRequest struct {
+		_ struct{} `route:"POST=/clone"`
+
+		CloneEnvironmentId object.ID `path:"environment"`
+
+		model.EnvironmentCreateInput `path:",inline" json:",inline"`
+	}
+
+	RouteCloneEnvironmentResponse = model.EnvironmentOutput
+)
+
+func (r *RouteCloneEnvironmentRequest) Validate() error {
+	return validateEnvironmentCreateInput(r.EnvironmentCreateInput)
+}
